修复多用户登录时通知问题

+++ src/main/java/life/gzhmt/forums/forums/controller/QuestionController.java
@@ -32,7 +32,7 @@ public class QuestionController {
         //已读功能
         User user = (User)request.getSession().getAttribute("user");
         if(user!=null) {
-            if (user.getId() == questionDTO.getCreator()) {
+            if ((int)user.getId()==(int)questionDTO.getCreator()) {
                 commentService.updateType(user.getId(), questionDTO.getId());
             }
         }
